Move identify module into liminal
Drop raw_devices entirely.
