+++ src/engine/strat_engine/engine.rs
@@ -18,10 +18,9 @@ use crate::{
             devlinks,
             dm::{get_dm, get_dm_init},
             keys::StratKeyActions,
-            liminal::LiminalDevices,
+            liminal::{find_all, LiminalDevices},
             names::{validate_name, KeyDescription},
             pool::StratPool,
-            raw_devices::find_all,
         },
         structures::Table,
         types::{CreateAction, DeleteAction, DevUuid, RenameAction, ReportType, SetUnlockAction},
