--- src/engine/strat_engine/raw_devices/mod.rs
@@ -1,7 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-mod identify;
-
-pub use self::identify::{find_all, identify_block_device, DeviceInfo, LuksInfo, StratisInfo};
